split resolver behavior ref support
